BACKEND/requirements.txt
flask-security-too
Flask-Cors
flask-restful
flask 
Flask-SQLAlchemy
setuptools